+++ tnnt_discordbot_cogs/__init__.py
@@ -4,5 +4,5 @@ Application init
 
 default_app_config = "tnnt_discordbot_cogs.apps.TnntDiscordbotCogsConfig"
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 __title__ = "TN-NT Discordbot Cogs"