+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.0] - 2021-12-15
+
+### Added
+
+- Rens, Hek, Dodixie and Perimeter to the overall market list
+- Order counts
+
+
 ## [0.1.0] - 2021-11-30
 
 ### Changed